requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
pillow>=10.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
pandas>=2.2.0
